mv encodec_print_usage and encodec_params_parse into common
