+++ examples/common.cpp
@@ -1,10 +1,54 @@
 #include <string>
+#include <thread>
 #include <vector>
 
 #define DR_WAV_IMPLEMENTATION
 #include "dr_wav.h"
 
+#include "common.h"
+
 #define SAMPLE_RATE 24000
+
+
+void encodec_print_usage(char ** argv, const encodec_params & params) {
+    fprintf(stderr, "usage: %s [options]\n", argv[0]);
+    fprintf(stderr, "\n");
+    fprintf(stderr, "options:\n");
+    fprintf(stderr, "  -h, --help            show this help message and exit\n");
+    fprintf(stderr, "  -t N, --threads N     number of threads to use during computation (default: %d)\n", params.n_threads);
+    fprintf(stderr, "  -m FNAME, --model FNAME\n");
+    fprintf(stderr, "                        model path (default: %s)\n", params.model_path.c_str());
+    fprintf(stderr, "  -i FNAME, --input FNAME\n");
+    fprintf(stderr, "                        original audio wav (default: %s)\n", params.original_audio_path.c_str());
+    fprintf(stderr, "  -o FNAME, --outwav FNAME\n");
+    fprintf(stderr, "                        output generated wav (default: %s)\n", params.dest_wav_path.c_str());
+    fprintf(stderr, "\n");
+}
+
+int encodec_params_parse(int argc, char ** argv, encodec_params & params) {
+    for (int i = 1; i < argc; i++) {
+        std::string arg = argv[i];
+
+        if (arg == "-t" || arg == "--threads") {
+            params.n_threads = std::stoi(argv[++i]);
+        } else if (arg == "-m" || arg == "--model") {
+            params.model_path = argv[++i];
+        } else if (arg == "-o" || arg == "--outwav") {
+            params.dest_wav_path = argv[++i];
+        } else if (arg == "-i" || arg == "--input") {
+            params.original_audio_path = argv[++i];
+        } else if (arg == "-h" || arg == "--help") {
+            encodec_print_usage(argv, params);
+            exit(0);
+        } else {
+            fprintf(stderr, "error: unknown argument: %s\n", arg.c_str());
+            encodec_print_usage(argv, params);
+            exit(0);
+        }
+    }
+
+    return 0;
+}
 
 bool read_wav_from_disk(std::string in_path, std::vector<float>& audio_arr) {
     uint32_t channels;
